# Reduce CPU and memory usage incurred by metrics