--- a/common/noop_metrics.go
+++ b/common/noop_metrics.go
@@ -26,17 +26,17 @@ func NewNoopMetrics() MetricHandle {
 
 type noopMetrics struct{}
 
-func (*noopMetrics) GCSReadBytesCount(_ context.Context, _ int64)                         {}
-func (*noopMetrics) GCSReaderCount(_ context.Context, _ int64, _ string)                  {}
-func (*noopMetrics) GCSRequestCount(_ context.Context, _ int64, _ string)                 {}
-func (*noopMetrics) GCSRequestLatency(_ context.Context, _ time.Duration, _ []MetricAttr) {}
-func (*noopMetrics) GCSReadCount(_ context.Context, _ int64, _ string)                    {}
-func (*noopMetrics) GCSDownloadBytesCount(_ context.Context, _ int64, _ string)           {}
+func (*noopMetrics) GCSReadBytesCount(_ context.Context, _ int64)                   {}
+func (*noopMetrics) GCSReaderCount(_ context.Context, _ int64, _ string)            {}
+func (*noopMetrics) GCSRequestCount(_ context.Context, _ int64, _ string)           {}
+func (*noopMetrics) GCSRequestLatency(_ context.Context, _ time.Duration, _ string) {}
+func (*noopMetrics) GCSReadCount(_ context.Context, _ int64, _ string)              {}
+func (*noopMetrics) GCSDownloadBytesCount(_ context.Context, _ int64, _ string)     {}
 
-func (*noopMetrics) OpsCount(_ context.Context, _ int64, _ string)            {}
-func (*noopMetrics) OpsLatency(_ context.Context, _ time.Duration, _ string)  {}
-func (*noopMetrics) OpsErrorCount(_ context.Context, _ int64, _ []MetricAttr) {}
+func (*noopMetrics) OpsCount(_ context.Context, _ int64, _ string)                  {}
+func (*noopMetrics) OpsLatency(_ context.Context, _ time.Duration, _ string)        {}
+func (*noopMetrics) OpsErrorCount(_ context.Context, _ int64, _ FSOpsErrorCategory) {}
 
-func (*noopMetrics) FileCacheReadCount(_ context.Context, _ int64, attrs CacheHitReadType)   {}
-func (*noopMetrics) FileCacheReadBytesCount(_ context.Context, _ int64, _ []MetricAttr)      {}
-func (*noopMetrics) FileCacheReadLatency(_ context.Context, _ time.Duration, _ []MetricAttr) {}
+func (*noopMetrics) FileCacheReadCount(_ context.Context, _ int64, attrs CacheHitReadType) {}
+func (*noopMetrics) FileCacheReadBytesCount(_ context.Context, _ int64, _ string)          {}
+func (*noopMetrics) FileCacheReadLatency(_ context.Context, _ time.Duration, _ string)     {}