… common/telemetry.go
… common/telemetry_test.go
… internal/fs/wrappers/monitoring.go
… internal/gcsx/file_cache_reader.go
… internal/monitor/bucket.go
… tools/integration_tests/monitoring/prom_test.go