Reduce metric CPU usage from ~10% for single threaded reads to ~4% and memory allocation from 10% to ~1%.

The `metric.WithAttributes` method is inefficient and is not recommended for performance-sensitive scenarios. This PR moves it to use `metric.WithAttributeSet` method that takes an attribute-set as input and doesn't have those performance concerns. To optimize it further, we use a cache that takes attribute-keys as input, and returns a corresponding attribute.Set object. This reduces memory-allocations by creating an attribute.Set object just once.

There are more improvements that could be performed and will be taken up in the subsequent PRs.

### Description
Before fix: http://shortn/_zNJRaWzRB4
After fix: http://shortn/_nPnS1bb5Vy
### Link to the issue in case of a bug fix.
b/424403628

Perf:

| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 551.34MiB/s  | 1.26MiB/s  |  74.65MiB/s  |  1.17MiB/s   |
|   PR   |  0.25MiB   | 540.25MiB/s  | 1.18MiB/s  |  77.54MiB/s  |  1.21MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3630.12MiB/s | 85.51MiB/s | 1530.37MiB/s |  82.58MiB/s  |
|   PR   | 48.828MiB  | 3629.67MiB/s | 85.17MiB/s | 1485.41MiB/s |  85.75MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3622.02MiB/s | 78.52MiB/s | 764.61MiB/s  |  39.29MiB/s  |
|   PR   | 976.562MiB | 3755.86MiB/s | 77.99MiB/s | 1129.05MiB/s |  38.71MiB/s  |

### Testing details
1. Manual - Verified
2. Unit tests - NA
3. Integration tests - Covered by existing integration tests.

### Any backward incompatible change? If so, please explain.